refactored git operations (now merged to one) into separate class, shell commands should be moved over as well
