+++ lib/thinx/builder.js
@@ -27,6 +27,8 @@ var Version = require("./version"); var v = new Version();
 var Sanitka = require("./sanitka"); var sanitka = new Sanitka();
 var Platform = require("./platform"); var _platform = new Platform();
 
+var Git = require("./git"); var git = new Git();
+
 var shellescape = require('shell-escape');
 
 module.exports = class Builder {
@@ -205,38 +207,6 @@ module.exports = class Builder {
 		return true;
 	}
 
-	gitFetchWithKeys(owner, command) {
-		var success = false;
-		let result;
-		console.log("No problem for Sources, re-try using SSH keys (TODO: should cleanup first)...");
-
-		// Feature envy?
-		var rsakey = require("./rsakey"); 
-		var RSAKey = new rsakey();
-		const key_paths = RSAKey.getKeyPathsForOwner(owner);
-
-		if (key_paths.count < 1) {
-			console.log("no_rsa_keys_found"); // todo: build or audit log
-			return false;
-		}
-		for (var kindex in key_paths) {
-			var gfpfx = "ssh-agent bash -c 'ssh-add " + app_config.ssh_keys + "/" + key_paths[kindex] + "; ".replace("//", "/");
-			console.log("git prefix", kindex, gfpfx);
-			let cmd = gfpfx + command + "'";
-			console.log("git command: " + cmd);
-			try {
-				result = exec.execSync(cmd);
-				console.log("[sources] git rsa clone result: ", result.toString());
-				success = true;
-				break;
-			} catch (e) {
-				console.log("git rsa clone error: " + e);
-				success = false;
-			}
-		}
-		return success;
-	}
-
 	run_build(build_id, owner, git, branch, udid, dryrun, notifiers, callback) {
 
 		if (!this.buildGuards(callback, owner, git, branch)) return;
@@ -334,7 +304,7 @@ module.exports = class Builder {
 				const git_success = fs.existsSync(BUILD_PATH + "/*");
 				console.log("Initial prefetch successful? : " + git_success);
 				if (git_success == false || git_success == []) {
-					if (!this.gitFetchWithKeys(owner, SHELL_FETCH)) {
+					if (!git.fetch(owner, SHELL_FETCH, BUILD_PATH)) {
 						console.log("Secondary git fetch failed. That's bad.");
 						console.log("Saving state FAILED for build_id", build_id, "owner", owner, "udid", udid);
 						blog.state(build_id, owner, udid, "failed");
