+++ lib/thinx/sources.js
@@ -19,6 +19,8 @@ var Platform = require("./platform"); var _platform = new Platform();
 
 var shellescape = require('shell-escape');
 
+var Git = require("./git"); var git = new Git();
+
 const { v1: uuidV1 } = require('uuid');
 
 module.exports = class Sources {
@@ -165,61 +167,6 @@ module.exports = class Sources {
 				}); // callback
 			}); // insert
 		}); // destroy
-	}
-
-	gitFetchWithKeys(owner, command) {
-		var success = false;
-		let result;
-		console.log("No problem for Sources, re-try using SSH keys (TODO: should cleanup first)...");
-		const key_paths = RSAKey.getKeyPathsForOwner(owner);
-		if (key_paths.count < 1) {
-			console.log("no_rsa_keys_found"); // todo: build or audit log
-			return false;
-		}
-		for (var kindex in key_paths) {
-			var gfpfx = "ssh-agent bash -c 'ssh-add " + app_config.ssh_keys + "/" + key_paths[kindex] + "; ".replace("//", "/");
-			console.log("git prefix", kindex, gfpfx);
-			let cmd = gfpfx + command + "'";
-			console.log("git command: " + cmd);
-			try {
-				result = exec.execSync(cmd);
-				console.log("[sources] git rsa clone result: ", result.toString());
-				success = true;
-				break;
-			} catch (e) {
-				console.log("git rsa clone error: " + e);
-				success = false;
-			}
-		}
-		return success;
-	}
-
-	gitPrefetch(CMD, local_path) {
-		var result = null;
-		try {
-			console.log("if running command will fail, needs catching the error as well... function should have callback\n: " + CMD);
-			result = exec.execSync(CMD);
-		} catch (e) {
-			console.log("[sources] clone_and_recurse_exception " + e);
-			// Should not exit until trying keys... no return here.
-		}
-
-		let rstring;
-
-		if (typeof (result) !== "undefined" && result !== null) {
-			rstring = result.toString();
-			console.log("GIT clone error: ", rstring);
-
-			if (rstring.indexOf("make sure you have the correct access rights") !== -1) {
-				console.log("Initial prefetch failed.");
-				return false;
-			}
-		}
-
-		// Try to solve access rights issue by using owner keys...
-		var git_success = fs.existsSync(local_path + "/*");
-		console.log("Initial prefetch successful? : " + git_success);
-		return git_success;
 	}
 
 	cleanupDirectory(cleanup_path) {
@@ -376,10 +323,10 @@ module.exports = class Sources {
 			"fi";
 
 		// Try to solve access rights issue by using owner keys...
-		const git_success = this.gitPrefetch(PREFETCH_CMD, TEMP_PATH);
+		const git_success = git.fetch(PREFETCH_CMD, TEMP_PATH, source.owner);
 		if (git_success == false || git_success == []) {
 			this.cleanupDirectory(TEMP_PATH);
-			if (!this.gitFetchWithKeys(source.owner, PREFETCH_CMD)) {
+			if (!git.fetch(source.owner, PREFETCH_CMD, TEMP_PATH)) {
 				callback(false, "Git fetch failed.");
 				return;
 			}
@@ -399,6 +346,7 @@ module.exports = class Sources {
 					source.initial_platform = platform; // should happen only on add
 					console.log("Will save source to owner: " + { source }, " path: " + TEMP_PATH + " with callback.");
 					this.addSourceToOwner(source.owner, source, TEMP_PATH, callback); // returns success, response
+					this.cleanupDirectory(TEMP_PATH);
 				} break;
 				case false: {
 					console.log("getPlatform failed! Platform: " + platform);
